Refactoring: Wrapper round <html>
To enable multiple border images on <html> (or at least giving that
effect).

The wrapper round <html> has to be <body>. So have #body2 doing what
<body> was doing before.

Except since <body> is called <body>, and not #html2, it looks odd to not
set things like the font there, leaving them to this weird <div>. So keep
those on <body> anyway.