Enhancement: Better scrollbar handling
Avoid having excess space just in case there's a particularly wide
scrollbar.

Add the relevant width to pages without a scrollbar, rather than
subtracting it from pages with one. That avoids needing to have an
explicit 'margin of error' added in the first place.

This reverts parts of 0f81f3d and 34201f8.